+++ tempest_fastapi_sdk/auth/__init__.py
@@ -35,6 +35,7 @@ from tempest_fastapi_sdk.auth.schemas import (
     LoginResponseSchema as LoginResponseSchema,
 )
 from tempest_fastapi_sdk.auth.schemas import LoginSchema as LoginSchema
+from tempest_fastapi_sdk.auth.schemas import LogoutSchema as LogoutSchema
 from tempest_fastapi_sdk.auth.schemas import MFAConfirmSchema as MFAConfirmSchema
 from tempest_fastapi_sdk.auth.schemas import MFADisableSchema as MFADisableSchema
 from tempest_fastapi_sdk.auth.schemas import (
@@ -68,6 +69,7 @@ __all__: list[str] = [
     "ActivationToken",
     "LoginResponseSchema",
     "LoginSchema",
+    "LogoutSchema",
     "MFAConfirmSchema",
     "MFADisableSchema",
     "MFAEnrollResponseSchema",